Move ControlsSettingsFragment to a new file
