+++ AnkiDroid/src/main/res/xml/preference_headers.xml
@@ -63,7 +63,7 @@
 
     <!-- Keybinding preferences -->
     <Preference
-        android:fragment="com.ichi2.anki.Preferences$ControlsSettingsFragment"
+        android:fragment="com.ichi2.anki.preferences.ControlsSettingsFragment"
         android:key="@string/pref_controls_screen_key"
         android:title="@string/pref_cat_controls"
         android:icon="@drawable/ic_touch">
